python-worker/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
insightface==0.7.3
onnxruntime>=1.17.0
open-clip-torch==2.24.0
transformers==4.36.0
torch>=2.2.0
chromadb==0.4.22
pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80
httpx==0.26.0
python-dotenv==1.0.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Fix for ChromaDB telemetry error
posthog<6.0.0
